+++ qunit/config.js
@@ -5,9 +5,9 @@ require.config({
 		"jqlite": "../node_modules/jqlite/jqlite",
 		"qunit": "../node_modules/qunitjs/qunit/qunit",
 		// "inputmask.dependencyLib": "../dist/inputmask/inputmask.dependencyLib.jquery",
-		"inputmask.dependencyLib": "../extra/dependencyLibs/inputmask.dependencyLib",
-		// "inputmask.dependencyLib": "../extra/dependencyLibs/inputmask.dependencyLib.jqlite",
-		"inputmask": "../dist/inputmask/inputmask"
-		// "inputmask": "../js/inputmask"
+		// "inputmask.dependencyLib": "../extra/dependencyLibs/inputmask.dependencyLib",
+		"inputmask.dependencyLib": "../extra/dependencyLibs/inputmask.dependencyLib.jqlite",
+		// "inputmask": "../dist/inputmask/inputmask"
+		"inputmask": "../js/inputmask"
 	}
 });