+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file.
 - include textarea as a valid masking element
 
 ### Updates
+- implement missing parts in the jqlite DependencyLib
+- Remove namespaces for events (simplifies implementing other dependencyLibs, besides jquery)
 - update alternation logic
 
 ### Fixed